Revert "Revert "Bg fix collecting all meetups 163165544""
