+++ app/api/v1/models/meetup_model.py
@@ -3,9 +3,9 @@ from ..utils.utils import generate_id
 
 meetups = []
 
-
 class Meetup(object):
     """ Model class for the meetup object """
+
     def save(self, data):
         """ Function to save new meetup """
         data['id'] = generate_id(meetups)
