+++ app/api/v1/views/meetup_view.py
@@ -1,52 +1,30 @@
-from flask import jsonify, request
-from ...v1 import version_1 as v1
-from ..schemas.meetup_schema import MeetupSchema
-from ..models.meetup_model import Meetup
+import os
+from flask import Flask, jsonify
+from instance.config import app_config
+from flask_jwt_extended import (JWTManager)
+from app.api.v1.views.meetup_view import v1 as meetups_blueprint_v1
 
 
-db = Meetup()
+def create_app(config_name):
+    """ Function to initialize Flask app """
 
+    config_name = os.environ.get('FLASK_CONFIG', 'development')
 
-@v1.route('/meetups', methods=['POST'])
-def create_meetup():
-    """ create meetup function """
-    json_data = request.get_json()
+    # Initialize app
+    app = Flask(__name__, instance_relative_config=True)
 
-    # when no data is provided
-    if not json_data:
-        return jsonify({'status': 400, 'error': 'No data provided'}), 400
+    app.config.from_object(app_config[config_name])
 
-    # whether the request is valid
-    data, errors = MeetupSchema().load(json_data)
-    if errors:
-        return jsonify({'status': 400, 'error' : 'Invalid data. Please fill all required fields', 'errors': errors}), 400
+    app.config.from_pyfile('config.py')
 
-    # Save new meetup when request is verified
-    new_meetup = db.save(data)
-    result = MeetupSchema().dump(new_meetup).data
-    return jsonify({'status': 201, 'message': 'Meetup created successfully', 'data': [result]}), 201
-
-
-
-@v1.route('/meetups/<int:meetup_id>', methods=['GET'])
-def fetch_meetup(meetup_id):
-
-    """  fetch specific meetup function """
-    # Check if meetup exists 
-    if not db.exists('id', meetup_id):
-        return  jsonify({'status': 404, 'error': 'Meetup not found'}), 404
-
-    # Get meetups 
-    meetups = db.fetch_by_id(meetup_id)
-    result = MeetupSchema(many=True).dump(meetups).data
-    return jsonify({'status':200, 'data':result}), 200
-
-
-@v1.route('/meetups/upcoming', methods=['GET'])
-def fetch_upcoming_meetups():
-    """  fetch all meetups fuction """
-    meetups = db.all()
-    result = MeetupSchema(many=True).dump(meetups).data
-    return jsonify({'status':200, 'data':result}), 200
-
+ 
+    # Register V1 Blueprints
+   
+    app.register_blueprint(meetups_blueprint_v1)
     
+   
+
+
+    return app
+
+
